[#10] Migrate @misk/core to Rush
* Update misk-web docker script to only run commands in directories that have a `package.json` present
