+++ misk-web/build.gradle
@@ -6,7 +6,7 @@ sourceSets {
   main.resources {
     srcDirs += [
       'web/packages/@misk/common/lib',
-      'web/@misk/core/lib',
+      'web/packages/@misk/core/lib',
       'web/packages/@misk/dev/lib',
       'web/packages/@misk/tslint/lib',
     ]
