+++ misk-web/web/packages/@misk/common/package.json
@@ -62,6 +62,7 @@
   "devDependencies": {
     "@misk/dev": "^0.1.1",
     "@misk/tslint": "^0.1.1",
+    "@types/history": "^4.7.2",
     "@types/lodash-es": "^4.17.1",
     "@types/node": "^10.12.18",
     "@types/react": "^16.7.18",
