+++ misk-web/web/packages/@misk/core/src/utilities/interfaces.ts
