+++ misk-web/web/packages/@misk/core/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../node_modules/@misk/dev/tsconfig.base",
+  "extends": "./node_modules/@misk/dev/tsconfig.base",
   "compilerOptions": {
     "outDir": "./lib/web/@misk/core"
   }
