+++ misk-web/web/packages/rush.json
@@ -257,6 +257,10 @@
       "projectFolder": "@misk/common"
     },
     {
+      "packageName": "@misk/core",
+      "projectFolder": "@misk/core"
+    },
+    {
       "packageName": "@misk/dev",
       "projectFolder": "@misk/dev"
     },
